+++ knishioclient/response/ResponseContinuId.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from ..models import Wallet
 from .Response import Response
+from .ResponseWalletList import _wallet_from_data
 
 
 class ResponseContinuId(Response):
@@ -10,4 +10,4 @@ class ResponseContinuId(Response):
             return None
 
         wallet_data = data[0] if isinstance(data, list) and len(data) > 0 else data
-        return Wallet.json_to_object(wallet_data)
+        return _wallet_from_data(wallet_data)